GlobalStructInference: Handle cases with just 1 global too (#5253)
Expand GlobalStructInference to operate on cases with a single possible global, and
not just 2 or more. Even the case of a single global is useful, it turns out, as we can
alter the reference in places like this:

(struct.get $type 0
  (..ref..)
)

No matter what ref is, if there is a single global it must refer to, we can switch to
this:

(struct.get $type 0
  (global.get $global)
)

That can unlock further opts later. Note that we can do this even if we don't know
what the value actually is - we may not know what the struct.get returns, but we
do know what it reads from.