[Wasm GC] Start an OptimizeCasts pass and reuse cast values there (#5263)
(some.operation
  (ref.cast .. (local.get $ref))
  (local.get $ref)
)

=>

(some.operation
  (local.tee $temp
    (ref.cast .. (local.get $ref))
  )
  (local.get $temp)
)

This can help cases where we cast for some reason but happen to not use the
cast value in all places. This occurs in j2wasm in itable calls sometimes: The
this pointer is is refined, but the itable may be done with an unrefined pointer,
which is less optimizable.

So far this is just inside basic blocks, but that is enough for the cast of itable
calls and other common patterns I see.