[NFC] Refactor GlobalTypeRewriter to split out the type mapping logic (#5295)
The logic that is split out into mapTypes gets a map of old type => new type and then
updates the module to replace the old with the new.

This will be useful in a future pass to merge types. We will tell it to map the types to
be merged with the types we want to merge them into.

This removes an assert on all types being in the map to allow some of them not to be.
(the new pass that will use this will only want to map some types).