Work around possible unsafe MSVC NaN optimization
MSVC 2019 is getting all parsed NaN signs backwards, even with `/fp:precise` or
`/fp:strict`. Those options only cause floats to conform to the source at
assignments, casts, and on function boundaries, but notably not in intermediate
results. My guess is that the double produced by the ternary statement is
considered an intermediate result, so MSVC considers itself free to perform
unsafe optimizations on it. Avoid the ternary statement to fix this problem if
that hypothesis is correct.