Fix Asyncify assertions after #5293 (#5328)
Followup to #5293, this fixes a small regression there regarding assertions. We do have
a need to visit non-instrumented functions if we want assertions, as we assert on some
things there, namely that such functions do not change the state (if they changed it,
we'd need to instrument them to handle that properly).

This moves that logic into a new pass. We run that pass when assertions are enabled.

Test diff basically undoes part the test diff from that earlier PR for that one file.