[Wasm GC] Add TypeMerging pass (#5321)
This finds types that can be merged into their super: types that add no
fields, and are not used in casts, etc. - so we might as well use the super.

This complements TypeSSA, in that it can merge back the new types that
TypeSSA created, if we never found a use for them. Without this, TypeSSA
can bloat binary size quite a lot (I see 10-20%).