Validate ref.as_* argument is a reference (#5330)
Without this we hit an assert with no line number info (or in a no-asserts build,
bad things can happen). With this:

$ bin/wasm-opt -all ~/Downloads/crash.wat  --nominal
[parse exception: Invalid ref for ref.as (at 155065:119)]
Fatal: error parsing wasm

(That can only happen for ref.as_non_null, as all the others do not have that
assert - their types do not depend on the child's type, so their finalize does not
error. Still, it is nice to validate earlier for them as well, so this PR handles them
all.)