Do not optimize public types
Do not optimize or modify public heap types in any way. Public heap types
include the types of imported or exported functions, tables, globals, etc. This
is important to maintain the public interface of a module to ensure it can still
link interact as intended with the outside world.

For now, error out if we find any public types that refer to nontrivial other
types. This slightly simplifies the code, but more importantly, it ensures that
all the type optimizations are not silently inhibited. In the future, we may
want to add options to silence this error or downgrade it to a warning.

This commit only updates the type updating machinery to avoid updating public
types. It does not update any optimization passes accordingly, so it's not
expected to work quite right yet.