wasm2js: Avoid emitting non-JS code during opt
As noted in #4806, trying to optimize past level 0 can result in
passes emitting non-JS code, which is then unable to be converted during
final output.

This commit creates a new targetJS option in PassOptions, which can
be checked inside each pass where non-JS code might be emitted.

This commit initially adds that logic to OptimizeInstructions, where
this issue was first noticed.