[Wasm GC] Enforce closed-world in GlobalStructInference (#5385)
After making the pass error when not in closed world, some testcases
required changes. One forward-looking testcase can just be removed - the
point of it was to see what happens if a type escapes, and now closed-world
errors on it - and another testcase needed to avoid types on the boundary,
and to use anyref+casts.