[Wasm GC] Do not merge supertypes into subtypes
In TypeMerging we previously merged all subsequent types in a refined partition
into whichever type happened to be first in the partition, but when that first
type happened to be a subtype of one of the other types in the partition, that
would cause type-updating.cpp to try to update that subtype's supertype to be
itself, causing an assertion failure.

Fix the problem by ensuring that the merge target is not a subtype of any other
types in the partition.