[NFC] Experiment with C++20 coroutines for expression traversal
Add a gtest file that is only included in the build when using C++20 or greater
that defines a generator utility in terms of C++20 coroutines and then uses that
to define generators for traversing Expressions. Unlike our current traversal
utilities, these generators leave control flow up to their callers, making it
trivial to do things like stop the traversal early or run multiple traversals
concurrently, for example to compare two expression trees.