RemoveUnusedModuleElements: Optimize unread struct.new fields (#5445)
This is similar to the existing optimization for function references: a RefFunc is
just a reference, and we do not consider the function actually used unless some
CallRef can actually call it. Here we optimize StructNew fields by tracking if they
are read. Consider this object:

(struct.new $object
  (global.get $vtable)
)

Before this PR, when we saw that StructNew we'd make that global used, and
process all of its contents, which look like this:

(global $vtable
  (struct.new $object
    (ref.func $foo)
  )
)

With this PR we track which fields on the struct types are even read. If they
are not then we do not add uses of the things they refer to. In this example,
the global vtable would be referenced, but not used, and likewise the RefFunc
inside it.

The technical way we handle this is to walk StructNews carefully: when we see
a field that has been read, we can just read it normally, but if it has not been
read then we note it on the side, and only process it later if we see a read.