[Wasm GC] Fix RemoveUnusedModuleEffects struct field reads with call.without.effects (#5477)
If we see

(struct.new $A
  (call $call.without.effects ;; intrinsic
    (ref.func $getter)
  )
)

then we can ignore side effects in that call, as the intrinsic tells us to. However,
that function is still called (it will be called normally, after intrinsics are lowered
away), and we should not remove it.

That is, such an intrinsic call can be removed, but it cannot be left as it is
and also ignored as if it does not exist.