[wasm-ctor-eval] Add support for multivalue serialization and a quiet mode (#5510)
Simply loop over the values and use tuple.make.

This also adds a lit test for ctor-eval. I found that the problem blocking us
before was the logging, which confuses the update script. As this test at
least does not require that logging, this PR adds a --quiet flag that
disables the logging, and then a lit test just works.