Parse and print `array.new_fixed` (#5527)
This is a (more) standard name for `array.init_static`. (The full upstream name
in the spec repo is `array.new_canon_fixed`, but I'm still hoping we can drop
`canon` from all the instruction names and it doesn't appear elsewhere in
Binaryen).

Update all the existing tests to use the new name and add a test specifically to
ensure the old name continues parsing.